+++ src/components/ResourcesEditor/ResourceEditor.styles.ts
@@ -16,10 +16,13 @@ export const getStyles = (theme: GrafanaTheme2) => {
       overflow: hidden;
       text-overflow: ellipsis;
     `,
+    dragHandle: css`
+      display: flex;
+      margin: ${theme.spacing(0, 0.5)};
+    `,
     dragIcon: css`
       cursor: grab;
       color: ${theme.colors.text.disabled};
-      margin: ${theme.spacing(0, 0.5)};
       &:hover {
         color: ${theme.colors.text};
       }